Initial implementation of wizard factories
In order to avoid getting wizard from pool (which forces to get session_id from create method and then instantiate
it), a new class called WizardFactory is created.

With this initial approach the user can define Wizard as a factory and not taking care of session. The purpose of
this class is lay the grundwork to allow manage ModelViews and state transistions in the future.